requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
Django==3.0.8
django-filter==2.3.0
django-graphql-geojson==0.1.4
fastdiff==0.2.0
graphene==2.1.8
graphene-django==2.12.1
graphql-core==2.3.2
graphql-relay==2.0.1
model-mommy==2.0.0
postgis==1.0.4
promise==2.3
psycopg2-binary==2.8.5
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
snapshottest==0.5.1
sqlparse==0.3.1
termcolor==1.1.0
Unidecode==1.1.1
wasmer==0.4.1
